move TopicsContentPanelIndex to correct place
